STERI016-86: Move new header code to nav.js and nav.css files
